Add error and not-found handling to the font glyph loader
When loading a given glyph, there could be errors; we'll just
return to the caller in that case.  Similarly, the font we're
using to find a given glyph may not even have it, so no sense
filling the glyph cache with non-data; we'll return to caller in
that case as well.