Move box-size font mechanism into base_font
When we do a box-size query to a font-list, we'll need to have
some sort of cached values, so we'll use the same scheme as our
single font object.  Cache the values in a few data members, and
return them.  No need to duplicate anything, so we'll move the
entire mechanism into the base class.