Add font_set object
The font_set will contain a set of fonts which we'll use to search
for a given glyph.  If the first doesn't have what we need, we'll
go down the list.

We'll use the injection operator to add fonts into the list.  It
has the benefit of not introducing any new call names, and still
gets the point of what we're doing across.  In order to inject a
single font at a time, without resorting to a bunch of hoops to
force things in a particular order, we'll inject tuples.  Each
tuple will have a name, a pixel size, and the search path.

The box size for a font set will be the box size of the first font
in the list.  We may want to revisit that at a later time, and
calculate the largest box of all the fonts in the set.

'set' may not be the right descriptive term for how this is
implemented, but it sounds more right for what we're after.

Re:  issue #1