Add a second font to the font test program
The true test of the operation of the font list is whether it
properly handles multiple fonts, in which we know that the first
font doesn't have some of the glyphs.  We add a new font, and
some small changes to the font test program to use the new font
and unicode range.

We also choose an R-to-L font here to improve coverage of those
codepaths which have gone previously unexercised.  Most R-to-L
scripts which are in current usage originated somewhere near or in
the middle east, so choosing one in today's politically-charged
climate will be seen by some as consorting with the enemy or
something.  To those people I say:  get over yourselves, it's just
text.

Re:  issue #1