Add checking for Freetype bitmap support in font
We're about to start in on supporting emojis, which requires
Freetype to be built with embedded bitmap support.  Fortunately
there is a header which already has that information, so we don't
need to add anything to our config.h file.  We'll just report the
findings, for people who are interested.

Re:  issue #2