Attempt to load a bitmap if there is no scaled glyph
If there is no scaled glyph in our font, FT_Load_Char will fail.
If that's the case, we'll call FT_Load_Glyph, which *should* load
bitmap data into the glyph slot.  The API docs say that color
bitmaps should be converted to greyscale if they are not loaded
with the FT_LOAD_COLOR load flag, but that is not actually
happening.