Repair thread-safety of strerror in cache
The cache used strerror to do some reporting of errors with its
reaper thread, which suggests that we should be using the thread-
safe variant straight away.

Also, we didn't have the right includes in the first place.

We also check in the autoconf to make sure we've got strerror_r.