Make label string rendering dependent on the string
We have a flag called 'use_text' in the label, which was intended
to let the fragment shader know whether it was getting one byte per
pixel or four.  We can get the same behaviour if we use the image
itself, which does note how many bytes each pixel gets, to set this
flag.

Once we start rendering text in full color, we'll continue to
apply the textures correctly, without having to make a breaking
change.  At that point, we can remove the flag, the uniform that
it modifies, and remove that handling from the fragment shaders.