Overlay cell structure on image/bitmap buffers
In order to make assigning RGBA values into image and bitmap
buffers, we'll overlay the notion of a "cell", which is four
unsigned ints in a row.  We always want to use RGBA, so we'll name
the pieces of the cell appropriately.

We'll eventually add some assignment operators, to make the usage
even more concise, but for now we just want to add the overlay
structure.