Add file with BIDI character classes
The next rule we need to implement is rule P2, which requires us
to detect things in classes AL and R.  For the first cut at doing
these classifications, we'll try using std::unordered_map objects
for the required classes.  The maps are kind of big, but they're
approximately constant-time access.