Add stubbed BIDI rule P2/P3 function
Rules P2 and P3 are mostly meant to operate together; they set the
base embedding (direction) of a paragraph based on the first non-
isolated character we find in the string.  The default is to use
an embedding of 0 if we can't find anything worthwhile, so we'll do
only that for our first version.