Query classes R and AL for characters for rules P2/P3
We have the classes as sets, so we can perform lookups for each
character we see.  Since the default is 0, and we'll always return
it when we don't find anything else of interest, we can bail out
when sensing an R or AL character.  The RLM and ALM marks act as
strong directions, so we can use them as valid entities for this
rule.

We don't yet handle directional isolates, which is a big part of
rule P2.