Observe directional isolates in rule P2
A large part of rule P2 is that we ignore any characters which are
inside directional isolates (denoted by LRI/RLI/FSI and PDI).  The
isolates can be nested, so we need to be able to roll and unroll
them as we enter and leave.  We'll only perform the legitimate test
when we're not in any isolate block.