Allow disabling of the GUI tests
The configure validates the usability of various perl modules,
including X11::GUITest.  On OSX, that starts up the XQuartz system,
which can take some time.  This change adds the --disable-gui-tests
configure flag, to disable such checking, and the running of the
GUI tests.  The uitest program can still be built by hand, if
desired.