Add BIDI character class enum
We will eventually need to classify all characters, so we'll add an
enum that contains all the character classes.  This will collide
with the existing set of defines, and turn them all weird, so we'll
change the single-character classes into const ints; we'll still be
able to use them the same, but there won't be any weird collision.