Create unicode_bidi class
We'll start inserting things into a class.  Once we start adding
the X rules in, we'll want to have some saved state available, so
we don't have to pass it all around.

We'll just move the things we have into the class, and worry about
moving state out of the function parameters and into the class
later on.