Add method to reset direction class for overridden classes
In several of the BIDI X rules, the class that we use for a given
character depends if we are in an override block.  If the top of
the direction stack has an override class in effect, it will force
the class of characters within the block to conform to that
override.