Add method for rule X2
Rule X2 is invoked when we see the RLE character in our string.
There is a specific set of circumstances that happen, from the spec
document.

The character_rec structure we use will eventually be used to keep
all the characters in our string, along with their embedding levels
and derived classes.  We don't yet do anything with them, but
eventually we'll put them into a container of some sort and use
them for the reordering operations.