Add method for rule X5c
Rule X5c handles FSIs in our input strings.

The FSI is used to create an isolate block whose direction is
keyed off the first strong character in the block.  We call out to
the P2/P3 rule, and depending on the embedding we get back, we may
call either the X5a or X5b rules.  In order to test this as simply
as possible, we'll mock out everything; this requires us to make
these methods virtual in the base bidi class.