Prepare X1 test for more test sections
Now that we have all of the sub-rules which comprise rule X1, we
can add the rest of the implementation for X1.  The test will have
more subtests, which in other tests we denote with the 'st'
variable; we'll get some of the changes out of the way before we
start in on any of the other stuff.