Add clause to rule N0 for preceding strong character
Rule N0, when a character between two matching brackets is found
which does not match the embedding direction, requires that we
search backward in the string until we find a strong character.  If
we do find one which also does not match the embedding direction,
the brackets get assigned to that class.