Add start-of-sequence clause to rule N0
If we have a strong character inside a bracket pair which has a
type which is opposite the embedding direction, and we don't find
another another character preceding the bracket pair which has
that same direction, we'll check the start-of-sequence marker.  If
it has that same opposite direction, we'll set the brackets to the
same type.