Add method for BD13
BD13 splits a string into "level runs", which are just sequences
of characters which share the same embedding level.

There are some parts of BD13 which we do not yet implement:  when
there is a level run which is interrupted by a directional isolate
sequence, the two pieces of that level should be appended together
for purposes of the W, N, and I rules.  This will be added on in
forthcoming commits, but for the time being, this will enable us
to get the rest of the rules committed, and a full process in
place.