Virtualize some of the bidi methods
We will need to make another mock class for rule X10, which will
encompass all of the W, N, and I rules.  They'll all need to be
virtual in order to make the mock class.

Also, we noticed that the destructors in both the original class
and all our mock classes were not virtual as they should be.