Add method for rule L3
Rule L3 involves possible reordering of combining marks, which our
renderer does not currently do anything with, so we don't need to
implement it.  We do make it virtual, so this class can be
subclassed by somebody who does want to implement something here.

The no-op test is strictly for test coverage.