Add character set for mirrored characters
There is a set of characters which should be considered mirrorable
when their embedding level is odd.  Many are bracket characters,
but there are a number of other glyphs which should be mirrored
when rendered.

Our renderer does not currently handle these, but once we have the
data available, we can start doing it.