Add |= operators for image cell
When we copy a glyph into an image, we may overlap with some data
that's already in the image, so we always use |=.  Our cell needs
to be able to handle both regular cells and float vectors.  In
order to reuse the struct |= operator, we'll turn the float vector
assignment operator into a constructor.