Add glyph method to copy into an image
Instead of having the copying inline in the renderer, we'll have
the glyph do its own copying.  It may be a grayscale glyph, in
which case we'll have to convert it to RGBA with the help of the
foreground color, or it may be mirrored, in which case we want to
copy it backwards.  It may already be colored, which will allow us
to copy it without any conversion.

squash into 930d0b061fd713e0f5d2e03287060e1d2160bb1e