Begin refactoring overly-complex get_string_size method
We had previously decided that the right way to return string size
values was a vector of ints.  That's overly complicated, and treats
the indices of that vector as magic constants, so we'll start
moving toward individual GLuint references for the output
parameters.  First step is to factor the guts out into the new
method, and leave a wrapper method with the old signature.