Shift all string rendering to use the BIDI implementation
The moment of truth.  We shift the render method to take vectors
of mirror_t, which is what the bidi object provides, and use the
bidi object to feed it.

We need to do a little more work, turning the multiline renderer
and the regular renderer into the same string, getting rid of the
multiline label, etc.