Repair segfaults in rule W2
We're seeing some segfaults in rule W2, where there is a character
of class EN in the first position.  We need to break out of the
inner loop when we're at the start.  We also need to consider that
our EN might be at the start of the run sequence.