Update rect::set_size to not cast away const
We take in a const void *, and then we've been casting it into
non-const other types; we shouldn't do that.  We'll retain the
const qualifier in all of the casts we do, to ensure we're not
doing anything untoward.

While we're at it, we'll convert our C-style casts into C++-style
reinterpret_casts.  Fortunately we can make them not too awful.

Re:  issue #42