Rework casting in set methods for widget-derived objects
Now that we're taking in const void * to all the set methods, we
should not cast away const when we cast things into the required
types for actual setting.

While we're at it, we'll convert the C-style plain casts into
C++-style reinterpret_casts.  It's the stylistically right thing to
do, and keeps the code as much in the C++ idiom as possible.

Re:  issue #42