Resolve bug in get_font in label
In redoing all the casting in our get/set methods, we found that
the get_font method in the label doesn't cast anything, so it only
sets the local void *, and nothing ever makes it back out to the
caller.  We'll do our typical reinterpret_cast of the input pointer
to a pointer to a pointer to a font object, and all should be well.

Additionally, since this is a const method, the things that come
from our object are also considered const.  We need to do a
const_cast to get rid of the constness for the return.