Rework the variable-args get and set methods
The C++ way of doing varargs functions/methods changed in C++11, to
use a templated function with an ellipsis typename as one of the
template arguments.  It allows us to get rid of the get_va and
set_va methods, in favor of using only get and set for all
operations.

Unfortunately, we can't rely on an inherited behaviour; templates
of this sort can not be virtualized or overridden.  All classes
which use the vararg behaviour must be updated to include their own
variable-argument template methods.  We've added the GET_VA and
SET_VA preprocessor macros to do this, which isn't ideal, but it
does work.  It's boilerplate, but at least we've simplified it as
much as we could.

Since we lose the inherited operations, we've got to update the
code and tests in one whack.  As a result, this diff is extremely
large.  Since we're in giant diff territory already, we might as
well update the docs while we're at it.