Allow disabling of the GUI tests
The configure validates the usability of various perl modules,
including X11::GUITest.  On OSX, that starts up the XQuartz system,
which can take some time, and doesn't work properly anyway.  This
change adds the --disable-gui-tests configure flag to disable such
checking, and the running of the GUI tests.  If the GLFW prereqs
are found, we'll still build the program, but we won't run it
during 'make check'.

Previous attempts at this change made the uitest program not work
properly, but we've verified that everything still looks fine with
the --disable-gui-tests flag set.