Convert all set methods to take direct values
The basic change we want to make is the ability to pass in direct
integer values, or at least not have to pass pointers to the things
we want to set.  Each set method, from the set() down to each
set_<resource>() method, now take types which correspond to the
type we expect.  A few of the set methods which could previously
take either an integer or a glm::ivec2, are now two methods.  It
allows us to use overloading to get to the right method, at the
expense of a bit more maintenance.

As a side-effect, we now get type checking on all set actions from
the compiler, since we expect actual types now.

Since we now have a bunch of overloaded names in each class, we
need to do a 'using <parent>::set' to pull in all of the parent's
methods, so we get all of the types to propagate down the class
hierarchy.

We're still working on the get methods, to see if we can use
references, rather than void * which we have to cast.  The issue
we're working from only requests set methods, but now that we
have type checking on the input side, we'd like to have it on the
output side as well.

Re:  issue #7