Improve setting font elements in ui::label and descendents
The fact that we've been taking all of our set() arguments as const
objects is not correct for the way that we use a font.  Instead of
casting away const at our first opportunity, we'll just get rid of
the const specifiers in all of the font-related set() methods.

Furthermore, when we pass a ui::font or ui::font_set pointer into
one of the set() methods which is expecting a ui::base_font
pointer, we get type problems.  I thought C++ was supposed to treat
pointers automatically as any of their parent classes, but that
doesn't seem to be the case.  We add wrapper methods in the
ui::label to handle the simple cast for us.

Re:  issue #62