Move compiler blocks into matrix block for Travis config
We were getting a weird collection of builds, not quite what we
wanted, so we'll be more explicit:  we want gcc and clang builds
on each of the platforms we're targetting.

Also, Xenial moves libGLU and its dev files into its own package,
so we'll make sure it's installed on the Xenial image.

Re:  issue #64