Expand matrix with single compiler elements in Travis config
When we add multiple compiler attributes to each of the os objects,
we don't get individual builds, but single builds with multiple
compilers tagged in the environment.  We want individual builds,
so we'll expand both the os objects to be two, each with a single
compiler.

Also, we're seeing failures on Xenial for a missing RandR dev
environment, so we'll make sure it's installed.

Re:  issue #64