Move required packages into the main package list
It's not clear whether having packages listed in the matrix at the
same time as packages listed outside of the matrix works properly.
Builds on the parts of the matrix that use the extras fail, so we
can assume that perhaps the items within the matrix are ignored.
Hopefully, in doing this, we don't make the older Trusty builds
fail, since they won't be able to find these new couple of
packages.

Re:  issue #64