Stop uploading artifacts that don't exist
Many moons ago, we used to take a snapshot of our uitest screen,
in order to do a little troubleshooting on why it wasn't working.
Now, we don't do the screenshot, and don't have anything to upload,
so there's no need to do uploads of things that don't exist.

If we need to do uploads in the future, we can dig this snippet out
of version control.