Add varargs constructor to pie menu
Since the pie menu is a "leaf" widget, and doesn't have any
descendents, we'll start with it for the VA-ification of the
constructors.

Basic form is to move all the initialization into a private method
called init() which takes a parent, and calling out to it in the
actual constructors.

I don't think we'll be able to have a macro that does our
constructor templates, but fortunately they're simple enough.

Re:  issue #6