Add varargs constructor to button
The button turns out to be another "leaf" widget, so we don't have
to go back and modify construction calls in anything else.  Same
deal as the rest of these conversions so far:  move the guts of
the constructor to the init() method, drop the width and height
args from the constructor, and add the varargs constructor.

Re:  issue #6