Add varargs constructor to label
Same changes as the rest of the varargs conversions:  move
existing constructor guts to the init() method, drop the width
and height args from the actual constructor, and add the varargs
one.

A few other widgets depend on the label, so there are changes to
the button and text field in with this commit, to get the
construction arguments correct.

Re:  issue #6