Add varargs constructor to UI context
The UI context is a bit of a weird bird as far as our widget
hierarchy goes, but the conversion is largely the same.  We don't
get a parent, because we *are* the ultimate parent, so those args
all go away, leaving us with nothing in the base constructor.

Re:  issue #6