Convert all get() methods to use typed pointers
Let's get some type-checking on the variables we pass into the
get() methods by using typed pointers instead of void pointers.  All
the casting to and from void * was pretty lame, and we lost all the
type-safety we're supposed to get out of C++ in the first place.

We may not be able to use references as requested.  Having tried
them, we ran into strange interactions with the varargs methods, in
that the first reference in the list would be the right variable,
but subsequent references would not.  Typed pointers are a bit less
magical than the references would be, and the big bonus is that
they work properly.

Re:  issue #61