Call color getters in uitest program
In order to test the color-getting mechanism, we'll add some calls
to the methods in our uitest program.

While we're at it, the name of the retrieval callback isn't quite
descriptive of what it does any longer, so we'll improve the name
to print_widget_resources.  This should be all of the resources
that the widget provides.

We'll add in other printers in the future, to test other widget
types' resource sets.