Test retrieving button resources in uitest program
We'll add more coverage of the various getting methods, this time
for some of the label and button resources.  String, image, font,
and the various states of the button.  We'll include a call-through
to the visible state, since there is a clause in the button's
method.

We'll also add some individual-element getters for size and
position, which weren't called in the other getter-tester function.