Start adding the armable class
When we started looking at adding a toggle button class, we found
that we needed the stuff out of the label for both buttons, but
that the toggle button needed some of the things that the regular
button provided, but not some of the other things.  Notably, we
needed the armed/active states in both, but the behaviour of the
regular button in its border-changing was not what we wanted in
the toggle button.

So the armable was born.  For now, it merely contains the armed
and active members, and does nothing with them.  This is the first
step on a refactoring path that will take a handful of commits.

Once we get a bit more of the contents of the button moved into
the armable class, we'll insert it into the hierarchy.  Then we'll
be able to derive from it for the toggle button.