Move part of state-setting from button to armable
We'll move the base part of the state-setting methods from the
button to the armable, because we need the basic actions, but can't
define the behaviour that each subclass needs.  The regular button
handles arm/active with changes to the border size, but we want to
do something different with the forthcoming toggle button.

The armable will add some pure-virtual methods, which subclasses
will need to implement, in order to define that arm/active
behaviour.