Move arm/active callbacks from button to armable
The logic for changing the state of the button-ish widgets will be
handled within the setter methods, so we can move the callbacks
which trigger these state changes into the armable.  Any subclass
will only have to implement the set_arm_state and set_active_state
methods to have the same enter/leave + down/up behaviour.