Update manpage formats to be more manpage-y
Some of the manpages we've looked at recently have the argument
types in the descriptions of each function, along with the argument
name in italics.  This conveys better information, so we'll adopt
it across all our manpages.

Also some small improvements regarding the types for the get() and
set() methods in the ui::rect.  We changed how those work, and
updated the overview doc, but neglected to update the manpage.