Add callback logic for the toggle button
We want default behaviour of the toggle button to, well, toggle the
checked state.  We need to rearrange the logic flow a little bit
so that we can retain the active/armed states and query them when
making the determination of when we actually need to toggle the
state.

Re:  issue #27