Add a space for the toggle checkbox
We want the checkbox to be square, so we'll add an extra image
height to the desired width of the widget.  We also want it to be
on the left side of the widget, and the label forces its image to
the upper-left, so we'll scoot the image to the upper-right.

Re:  issue #27