Draw the actual checkbox and arm border for the toggle
Now that everything is in place, and states are correct, we'll draw
the checkbox, either empty or filled, and the arm border.

Since the vertex buffer object is statically-sized, it is simpler
to expand the size of the single vertex buffer that we'll get back,
instead of making a new one of the right size for us, and copying
everything into it.  Thus the modification to widget.cc.

Re:  issue #27