Convert vertex buffer to use std::vector internally
Since the vector is basically an array internally, we can make our
vertex buffer use it with very little change.  Each method which
adds things to the buffer now inserts more elements at the end
before it does anything, and the data-retrieval methods now return
the data() method of the vector.

This allows us to get rid of the object's static initialization
parameters.

Re:  issue #71