Add a method to retrieve the element count from the vertex buffer
We want to get rid of the vertex_index and element_index members,
because they are not really needed any longer, but we need a simple
method to retrieve the number of items in the element vector.  For
now we'll just use the element_index member, but this will change
soon.

Re:  issue #71