Remove element_index member from vertex buffer
We don't need to keep this data around, because we have easier ways
to get it - the element vector keeps track of its size so we don't
have to.  This allows us to pull a lot of the usage of that member
local to each method that needs it.

Re:  issue #71