Clean up configure checks for manpage generation
Instead of making --with-ronn the flag for making manpages, we'll
make --with-manpages the flag, and make --with-ronn a secondary
flag that specifies a different ronn path.

We'll try some testing as to whether the ronn program we find is
new enough to include some of the features that we want, i.e.
table support (first appeared in ronn-ng 0.7.5).  We're adding a
new m4 macro, ax_compare_version to perform the version comparison.

Re:  issue #18