Add focused member to composite
We'll have each composite widget keep track of which of its
children currently has keyboard focus.  The key callback method
changes the most, as it now uses the focused data member to direct
keyboard events, rather than the widget under the mouse cursor.

The mouse-down callback changes somewhat, to add in focusing
behaviour.  We only change focus if we need to change focus (i.e.
if we click the widget that's already focused, we don't do
anything).  We still pass the clicks through to the appropriate
child widget.

We also add a focus callback method in the composite, to handle
unfocus events coming from parents.

Re:  issue #4