Factor text field key-application into its own method
In order to handle the key repeat, we'll use the timeout mechanism.
To make the timeouts work, we'll need to perform the same key
application on every timeout expiration, so instead of repeating
the same code block, we'll make it into a method that we can call
where necessary.

Re:  issue #74