Set up key repeat timeout
The key repeat timeout mechanism is pretty simple:  if the key
we're repeating is still the one that's held down, do the repeat
and set up for the next one.  If the key is released, get rid of
the timeouts.

Since we do save a pointer to something for the client data in a
timeout, we'll keep things simple and just copy the existing key
call data structure.  Once the key comes back up, we'll make sure
to delete what we've allocated.

Re:  issue #74