Add a mutex to ensure key repeats stop correctly
In our testing, we're seeing key repeats going on forever.  If the
timeout is hitting while the key-up is triggered, it's even odds as
to who wins.

We'll ensure that the key-up method takes primacy over the timeout.
The key-up will make a hard lock, where the timeout will try a lock
and bail out if it can't lock the mutex.  This way, key-up will
always occur, even if it has to wait a tiny bit, and timeout will
happen if it can.

We also noticed that we're being a little too stringent in our
checking for "is this the right timeout argument" - if there is an
arg set, we'll treat it as good enough.  If two keys are being held
down, and one is released, all repeats should be dropped.  That's
reasonable behaviour for a text field anyway; holding down two keys
doesn't make a whole lot of sense in the first place.

Re:  issue #74