Expand ui::key values
We were using mostly what GLFW has provided, in terms of a key
callback which handles all key-down events, and a char callback
which handles key-down events for printing characters.  However, we
weren't getting all the events we needed into the text field, so we
had to extend the ui::key namespace to include ALL the keys.  This
has been coming for a long time, and it appears that now is the
time.

For the time being, we've taken almost all of the list that GLFW
provides; there may be more or less for other platform-agnostic
OpenGL frameworks.

This requires us to modify the key callback in the connector such
that we always assume that we get a key which exists, since all of
them do.  The regular key callback connector will only trigger on
non-printing characters, which we have placed at the end, after a
sentinel value.  We'll use the sentinel value to be our test for
whether we trigger or not.