Add focus and close callbacks to GLFW connector
X11 has a couple mechanisms surrounding autorepeat of characters.
The first is a global setting for the entire X server, handled via
XAutoRepeatOff/XAutoRepeatOn.  The second is a per-client setting
which delivers a series of key-down events while the key is held,
and a single key-up event when the key is released.  This mechanism
is governed by XkbSetDetectableAutoRepeat.

We've tried both, and unfortunately the server-global autorepeat
setting is the only one that really works for us.  We can make the
setting we need on window focus, and restore the original setting
when we leave the window, or when the window closes.

Since this is controlled within the connector, and not from any
setting within the CuddlyGL toolkit itself, we don't need to make
any changes to the config.h file, as it will not be available when
other programs may be built.  We can set a build-time define of
USING_X11 to turn these features on.

Re:  issue #74