Add a disconnect function to the GLFW connector
When closing a GLFW window via window decorations, i.e. via the
mechanisms supplied by the OS, the window close callback works
fine.  However, when using a hotkey to close the window, the close
callbacks are not called.  This seems to be by design, which is
disappointing.  The assumption is apparently that the callbacks and
the rest of the program will be operating within the same scope,
which is not the case here; we want programs which use our toolkit
to be able to have as little interaction with GLFW (or whichever
OpenGL windowing library) as possible.

To that end, we have to create a ui_disconnect_glfw function, to
handle the case where we quit via a hotkey handled via the various
key callbacks.