Improve manpage generation
The ronn program has a *lot* of command line options, particularly
when operating as a pipe.  When a single one of those options is
missing, the manpages won't be properly formatted, and will look
terrible.  So we'll add all the required flags to our suffix rule,
and manpages now look great.

While we're at it, we'll have git ignore the generated manpage
files.

Re:  issue #18