Install manpages with expected filenames
Make is not properly able to handle filenames with colons in them,
and we want to have our manpages named as the classes they
document, colons and all.

To that end, we'll treat the manpage builds as plain data which
will not be installed, and set custom install/uninstall rules for
the doc directory.  It's super hacky, but it is what's required to
get make to do what we want.

Re:  issue #18