Add radio_box resource to composite
We'll use the container of toggles as the arbiter of radio state,
so we have to have a resource that turns on radio-box arbitration.
We don't have a boolean state set of get/set methods yet, so we
have to add those as well.

Adding it to the composite instead of the manager (which already
has boolean state methods) is so the context can also function as
a radio manager.  There's no reason anybody *should* do that, but
we won't limit people's questionable choices.

Re:  issue #28