Add radio button behaviour to toggle and composite
When a toggle is selected which is parented by a radio box, all
other toggles in the radio box should be unselected.

We add a new resource element and group for this, which should
hopefully be extensible for other items we may need in the future:
ui::element::child.  The only type within the ui::child space is
ui::child::radio at the moment, but it may prove useful for other
items.

The two resources which govern radio behaviour are not ideal.  They
do not seem to be very cohesive with the rest of the resource set.
But they work for the time being.  If we come up with something
better, we should definitely change them.

Re:  issue #28