Factor composite child set method into multiple methods
The rest of our get/set methods divide the responsibility of
determining whether the right flags have been passed in into two
(or more) methods, and we dropped the entirety of our child setter
into the top-level set method.

We'll move most of the guts of it into a separate set_radio_child
method, to keep with the same pattern that we use everywhere else.