Add radio child get method to composite
We've currently got no way to figure out what in a radio box is
actually checked, so we'll add a get method and get_radio_child
sub-method to retrieve this info.

Since we don't retain any sort of index data or separate "which
radio is this" data anywhere, we'll leave it up to the caller to
make sense of what radio means what.  We don't know if they're
using strings or images for their radios, so we can't force any
specific meaning.

We'll also exercise the new stuff in a test.

Re:  issue #28