Escape literal * characters in manpages
We're running into some weirdness with the way markdown is trying
to format function prototypes which have pointers in them.  The
markdown parser is trying to interpret the literal '*' characters
as part of markdown, which is not what we want.  We can do a
backslash-escape of those characters, which should get us rendering
as we should.