Make sure manpage docs are distributed
Automake uses some standard prefixes and target names for its
files.  Our Markdown manpage docs should be installed as docs,
regardless of whether they are being processed by ronn into legit
manpages.  We need the list of Markdown docs in a variable, so we
can use it for both the list of files to be installed, and also for
the list of manpages to be created.