Eliminate vector arguments from font max_cell_size
Our get_string_size used to use a std::vector<int> to return its
sizes, but we realized that was overly complicated, and led to
magical constants in using the elements of the vector.  The
max_cell_size method has the same problem, so we'll make two
methods which provide either width/height or
width/ascender/descender, to mirror the get_string_size methods.

The only caller of this method is the ui::text_box, so we only have
to fix the callers in a couple small places. Our goal of making the
calling code clearer is achieved by this change.