Save relative and absolute positions in widget
Since our relative position may contain negative values, and there
are quite a number of other classes which would need to perform
calculations to get the absolute position, we'll save both the
relative and absolute positions at all times.  When queried, we
will always return the relative position.  This will cause trouble
for other classes which need the absolute position, but we will
remedy that in a future commit.

We'll also add a method to convert the relative position into the
absolute.

Re:  issue #81