Allow querying of relative or absolute position
Now that we save both relative and absolute position in each
widget, we need to be able to retrieve either of them.  Up to now,
we've only provided relative position, but things like composites
need to deal with absolute position.

We'll add a new ui::position element:  ui::position::absolute.
We have gotten extremely lucky in that the new element corresponds
to the next bit in a potential bitmask, so we can now treat the
position type as an actual bitmask.

To that end, the composite and quadtree both need to query the
absolute positions for their event-handling requirements.  We'll
add the ui::position::absolute bit into all their calls into the
various child widgets.

Re:  issue #81