Resolve ambiguous inheritance in manager
Since the manager inherits from both the composite and the widget,
there can be some ambiguous inheritance.  There have been other
instances in which we've needed to resolve such inheritances,
since the widget comes into the hierarchy from two directions.

Our new set method for plain ints is just such a case.  Previously
we only had the one from the rect, but now we've got one from the
widget, which is not inherited virtually by the manager.

Re:  issue #81