Correct numerical errors in font string sizer
When computing the size of a rendered string, we were running into
occasional numerical problems due to treating everything as
unsigned integers.

Fonts often have negative numbers as part of glyphs.  Since we
want the results as unsigned, we've been treating each of the
elements as unsigned, which causes a std::max operation to treat
each of the negatives we run into as HUGE.  We'll clamp any
negative values to 0 for both the ascender and descender, so we'll
get image sizes that make sense.