Add description of widget relative positioning
Widget objects may have positive or negative values for their
positions, which indicate offsets from either the right or bottom
of the parent.

The summary document and the label manpages both needed updates
for the types accepted by the position type.

Also repaired a small error.  In the resource table for position,
ui::position::x was used twice in both the widget and label pages.