Handle keyboard focus in text fields
Keyboard focus and mouse movement are now divided into two separate
sets of functionality.  We'll repurpose the enter and leave
callbacks in the text field for a new combined focus callback.

This should be all the changes we need to fully support the new
keyboard focus mechanisms.  Everything else will be handled at the
composite level.

We also update the uitest runner to click on the text field instead
of just moving into it.  We also click the window once we're ready
to exit, so keyboard focus is in the top-level.

Re:  issue #4